Rewrite qr MVVM module
