+++ AdyenActions/UI/View Controllers/QR Code/Models/QRCodeFlowType.swift
@@ -1,0 +1,12 @@
+//
+// Copyright (c) 2025 Adyen N.V.
+//
+// This file is open source and available under the MIT license. See the LICENSE file for more info.
+//
+
+import Foundation
+
+internal enum QRCodeFlowType {
+    case copyCode
+    case saveAsImage
+}
